+++ appyx-interactions/common/src/commonMain/kotlin/com/bumble/appyx/interactions/core/ui/context/UiContext.kt
@@ -1,6 +1,5 @@
 package com.bumble.appyx.interactions.core.ui.context
 
-import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.runtime.Immutable
 import kotlinx.coroutines.CoroutineScope
 
@@ -8,6 +7,5 @@ import kotlinx.coroutines.CoroutineScope
 data class UiContext(
     val coroutineScope: CoroutineScope,
     val transitionBounds: TransitionBounds,
-    val boxScope: BoxScope,
     val clipToBounds: Boolean
 )