+++ demos/appyx-interactions/desktop/src/desktopMain/kotlin/com/bumble/appyx/interactions/widgets/ui/TargetUiState.kt
@@ -47,7 +47,7 @@ class TargetUiState(
         return MutableUiState(
             uiContext = uiContext,
             rotationX = RotationX(
-                uiContext = uiContext,
+                coroutineScope = uiContext.coroutineScope,
                 target = rotationX,
                 displacement = scrollX.mapState(uiContext.coroutineScope) {
                     2.5f * clamp(-it, 0f, 1f)
@@ -55,7 +55,7 @@ class TargetUiState(
                 origin = TransformOrigin(0.075f, 0f),
             ),
             position = PositionOutside(
-                uiContext = uiContext,
+                coroutineScope = uiContext.coroutineScope,
                 target = position,
                 displacement = scrollX.mapState(uiContext.coroutineScope) {
                     val factor = 0.075f + smoothstep(0f, 1f, it)
@@ -68,19 +68,19 @@ class TargetUiState(
                 }
             ),
             scale = Scale(
-                uiContext = uiContext,
+                coroutineScope = uiContext.coroutineScope,
                 target = scale,
                 displacement = scrollX.mapState(uiContext.coroutineScope) { -0.1f * it },
             ),
             alpha = Alpha(
-                uiContext = uiContext,
+                coroutineScope = uiContext.coroutineScope,
                 target = alpha,
                 displacement = scrollX.mapState(uiContext.coroutineScope) {
                     clamp(it, 0f, 1f) + clamp(-it - itemsInStack, 0f, 1f)
                 },
             ),
             zIndex = ZIndex(
-                uiContext = uiContext,
+                coroutineScope = uiContext.coroutineScope,
                 target = zIndex,
                 displacement = scrollX.mapState(uiContext.coroutineScope) { -it }
             ),